# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "NABARUN API",
    "version": "2.0"
  },
  "servers": [
    {
      "url": "https://ngonabarun-stage.appspot.com",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/user/{id}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Retrieve user details by user id",
        "description": "Authorities : hasAuthority('SCOPE_read:user')",
        "operationId": "getUserDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ID",
              "enum": [
                "EMAIL",
                "AUTH_USER_ID",
                "ID"
              ]
            }
          },
          {
            "name": "Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseUserDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "nabarun_auth": []
          },
          {
            "nabarun_auth_apikey": []
          }
        ]
      }
    },
...
    "/api/common/data/referenceData": {
      "get": {
        "tags": [
          "common-controller"
        ],
        "summary": "Retrieve reference data with filter",
        "operationId": "getReferenceData",
        "parameters": [
          {
            "name": "names",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/RefDataType"
              }
            }
          },
          {
            "name": "donationType",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/DonationType"
            }
          },
          {
            "name": "currentDonationStatus",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/DonationStatus"
            }
          },
          {
            "name": "countryCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stateCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workflowType",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RequestType"
            }
          },
          {
            "name": "Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseMapStringListKeyValue"
                }
              }
            }
          }
        },
        "security": [
          {
            "nabarun_auth": []
          },
          {
            "nabarun_auth_apikey": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserAddress": {
        "type": "object",
        "properties": {
          "ref": {
            "type": "string"
          },
          "addressType": {
            "type": "string",
            "enum": [
              "PRESENT",
              "PERMANENT"
            ]
          },
          "addressLine1": {
            "type": "string"
          },
          "addressLine2": {
            "type": "string"
          },
          "addressLine3": {
            "type": "string"
          },
          "hometown": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        }
      },
      "UserDetail": {
        "type": "object",
        "properties": {
          "presentAndPermanentAddressSame": {
            "type": "boolean"
          },
          "profileCompleted": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "initials": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "gender": {
            "type": "string"
          },
          "about": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "pictureBase64": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRole"
            }
          },
          "roleString": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "primaryNumber": {
            "type": "string"
          },
          "phones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserPhoneNumber"
            }
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserAddress"
            }
          },
          "socialMediaLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSocialMedia"
            }
          },
          "memberSince": {
            "type": "string",
            "format": "date-time"
          },
          "activeContributor": {
            "type": "string"
          },
          "publicProfile": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "DELETED",
              "BLOCKED"
            ]
          },
          "loginMethod": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "PASSWORD",
                "EMAIL",
                "SMS"
              ]
            }
          },
          "attributes": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          }
        }
      },
      "UserPhoneNumber": {
        "type": "object",
        "properties": {
          "ref": {
            "type": "string"
          },
          "phoneType": {
            "type": "string",
            "enum": [
              "PRIMARY",
              "ALTERNATIVE"
            ]
          },
          "phoneCode": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "displayNumber": {
            "type": "string"
          }
        }
      },
      "UserRole": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string"
          },
          "roleName": {
            "type": "string"
          },
          "roleCode": {
            "type": "string",
            "enum": [
              "MEMBER",
              "CASHIER",
              "ASSISTANT_CASHIER",
              "TREASURER",
              "GROUP_COORDINATOR",
              "ASST_GROUP_COORDINATOR",
              "SECRETARY",
              "ASST_SECRETARY",
              "COMMUNITY_MANAGER",
              "ASST_COMMUNITY_MANAGER",
              "PRESIDENT",
              "VICE_PRESIDENT",
              "TECHNICAL_SPECIALIST"
            ]
          },
          "description": {
            "type": "string"
          }
        }
      },
      "UserSocialMedia": {
        "type": "object",
        "properties": {
          "ref": {
            "type": "string"
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "FACEBOOK",
              "INSTAGRAM",
              "WHATSAPP",
              "TWITTER",
              "LINKEDIN"
            ]
          },
          "mediaName": {
            "type": "string"
          },
          "mediaIcon": {
            "type": "string"
          },
          "mediaLink": {
            "type": "string"
          }
        }
      },
      "SuccessResponseUserDetail": {
        "type": "object",
        "properties": {
          "info": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "traceId": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "version": {
            "type": "string"
          },
          "responsePayload": {
            "$ref": "#/components/schemas/UserDetail"
          }
        }
      },
...
      "RequestType": {
        "type": "string",
        "enum": [
          "JOIN_REQUEST",
          "JOIN_REQUEST_USER",
          "CHECK_PAYMENT_GUEST",
          "CHECK_PAYMENT",
          "COLLECT_CASH_PAYMENT",
          "TERMINATION_REQUEST",
          "DONATION_PAUSE_REQUEST",
          "PROFILE_UPDATE_REQUEST"
        ]
      },
...
      "DonationStatus": {
        "type": "string",
        "enum": [
          "RAISED",
          "PAID",
          "PENDING",
          "PAYMENT_FAILED",
          "PAY_LATER",
          "CANCELLED",
          "UPDATE_MISTAKE"
        ]
      },
      "DonationType": {
        "type": "string",
        "enum": [
          "REGULAR",
          "ONETIME"
        ]
      },
...
      "KeyValue": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "displayValue": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "RefDataType": {
        "type": "string",
        "enum": [
          "USER",
          "DONATION",
          "PUBLIC",
          "ACCOUNT",
          "WORKFLOW",
          "ADMIN"
        ]
      },
      "SuccessResponseMapStringListKeyValue": {
        "type": "object",
        "properties": {
          "info": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "traceId": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "version": {
            "type": "string"
          },
          "responsePayload": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/KeyValue"
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "nabarun_auth": {
        "type": "http",
        "in": "header",
        "scheme": "bearer"
      },
      "nabarun_auth_apikey": {
        "type": "apiKey",
        "name": "X-Api-Key",
        "in": "header",
        "scheme": "apikey"
      }
    }
  }
}